Updated the gist filter to allow setting of the cache directory in _config.yml. The config option is gist_cache, e.g. gist_cache: ./cache/gist

+++ lib/jekyll/filters.rb
@@ -60,21 +60,23 @@ module Jekyll
     end
 
     def gist(id, file='')
+      cache_dir = @context.registers[:site].config['gist_cache']
+
       # If the cache folder exists then attempt to find a cached version.
-      if File.directory?('_gist_cache')
+      if cache_dir && File.directory?(cache_dir)
         cache = File.join(
-          '_gist_cache', "#{id}_#{file.gsub(/[^A-Za-z0-9\._]/, '_')}")
+          cache_dir, "#{id}_#{file.gsub(/[^A-Za-z0-9\._]/, '_')}")
         js = open(cache).read if File.exists?(cache)
       end
       # If we have no cached version get the normal one.
-      js ||= open("http://gist.github.com/#{id}.js?file=#{CGI.escape(file)}").read
+      js ||= open(
+        "http://gist.github.com/#{id}.js?file=#{CGI.escape(file)}").read
 
       # If we don't arleady have a cached and the directory exists create one.
-      if File.directory?('_gist_cache') && !File.exists?(cache)
-        File.open(cache, 'w') { |f| f.write(js) }
-      end
+      File.open(cache, 'w') { |f| f.write(js) } if cache && !File.exists?(cache)
 
-      js.match(/document.write\('(<div.+)'\)/)[1].gsub(/\\"/, '"').gsub(/\\\//, '/').gsub(/\\n/, '')
+      js = js.match(/document.write\('(<div.+)'\)/)[1]
+      js.gsub(/\\"/, '"').gsub(/\\\//, '/').gsub(/\\n/, '')
     end
   end
 end
